Added SumWithoutUnluckyOneThree.java
